server basics working localhost

+++ CinemaServer/src/main/java/manager/CategorieRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
  * Created by Quentin on 03/11/2017.
  */
 
-public interface CategorieRepository extends CrudRepository <Categorie, Long> {
+public interface CategorieRepository extends CrudRepository <Categorie, Integer> {
 }
