+++ CinemaServer/src/main/java/manager/FilmRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
  * Created by Quentin on 03/11/2017.
  */
 
-public interface FilmRepository extends CrudRepository <Film, Long> {
+public interface FilmRepository extends CrudRepository <Film, Integer> {
 }
