+++ CinemaServer/src/main/java/manager/PersonnageRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
  * Created by Quentin on 03/11/2017.
  */
 
-public interface PersonnageRepository extends CrudRepository <Personnage, Long> {
+public interface PersonnageRepository extends CrudRepository <Personnage, Integer> {
 }
