+++ CinemaServer/src/main/java/manager/RealisateurRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.repository.CrudRepository;
  * Created by Quentin on 03/11/2017.
  */
 
-public interface RealisateurRepository extends CrudRepository <Realisateur, Long> {
+public interface RealisateurRepository extends CrudRepository <Realisateur, Integer> {
 }
